+++ clients/admin-ui/src/types/api/models/PrivacyNoticeUpdate.ts
@@ -7,6 +7,7 @@ import type { EnforcementLevel } from "./EnforcementLevel";
 import type { GPPFieldMappingCreate } from "./GPPFieldMappingCreate";
 import type { NoticeTranslation } from "./NoticeTranslation";
 import type { PrivacyNoticeFramework } from "./PrivacyNoticeFramework";
+import type { MinimalPrivacyNotice } from "./MinimalPrivacyNotice";
 
 /**
  * Overriding Privacy Notice schema for updates - translations must be supplied or they will
@@ -26,4 +27,5 @@ export type PrivacyNoticeUpdate = {
   framework?: PrivacyNoticeFramework | null;
   gpp_field_mapping?: Array<GPPFieldMappingCreate> | null;
   translations: Array<NoticeTranslation>;
+  children: Array<MinimalPrivacyNotice> | null;
 };