+++ clients/privacy-center/constants/index.ts
@@ -1,4 +1,5 @@
 /* eslint-disable import/prefer-default-export */
-import { IdentityInputs } from "~/types/config";
+
+import { IdentityInputs } from "~/types/api";
 
 export const defaultIdentityInput: IdentityInputs = { email: "optional" };