+++ clients/privacy-center/types/api/models/ActionType.ts
@@ -1,0 +1,13 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+/**
+ * The purpose of a particular privacy request
+ */
+export enum ActionType {
+  ACCESS = "access",
+  CONSENT = "consent",
+  ERASURE = "erasure",
+  UPDATE = "update",
+}