+++ clients/privacy-center/types/api/models/AsyncStrategy.ts
@@ -1,0 +1,11 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+/**
+ * Async strategies: supports callback for now, but could be
+ * extended to polling as well in the future
+ */
+export enum AsyncStrategy {
+  CALLBACK = "callback",
+}