+++ clients/privacy-center/types/api/models/CloudConfig.ts
@@ -1,0 +1,8 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+export type CloudConfig = {
+  privacy_center_url?: string | null;
+  domain_verification_records?: Array<string>;
+};