+++ clients/privacy-center/types/api/models/DrpMeta.ts
@@ -1,0 +1,10 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+/**
+ * Enum to hold Drp metadata. Only version is supported at this time
+ */
+export type DrpMeta = {
+  version: string;
+};