+++ clients/privacy-center/types/api/models/Page_Dataset_.ts
@@ -1,0 +1,13 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+import type { Dataset } from "./Dataset";
+
+export type Page_Dataset_ = {
+  items: Array<Dataset>;
+  total: number | null;
+  page: number | null;
+  size: number | null;
+  pages?: number | null;
+};