+++ clients/privacy-center/types/api/models/Page_PolicyWebhookResponse_.ts
@@ -1,0 +1,13 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+import type { PolicyWebhookResponse } from "./PolicyWebhookResponse";
+
+export type Page_PolicyWebhookResponse_ = {
+  items: Array<PolicyWebhookResponse>;
+  total: number | null;
+  page: number | null;
+  size: number | null;
+  pages?: number | null;
+};