+++ clients/privacy-center/types/api/models/PrivacyNoticeDisabledResponse.ts
@@ -1,0 +1,11 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+/**
+ * A minimal response for updating just whether a Privacy Notice is disabled
+ */
+export type PrivacyNoticeDisabledResponse = {
+  disabled: boolean;
+  id: string;
+};