+++ clients/privacy-center/types/api/models/StatusEnum.ts
@@ -1,0 +1,11 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+/**
+ * The model for possible evaluation results.
+ */
+export enum StatusEnum {
+  FAIL = "FAIL",
+  PASS = "PASS",
+}