+++ clients/privacy-center/types/api/models/StorageApplicationConfig.ts
@@ -1,0 +1,9 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+import type { StorageTypeApiAccepted } from "./StorageTypeApiAccepted";
+
+export type StorageApplicationConfig = {
+  active_default_storage_type: StorageTypeApiAccepted;
+};